+++ src/serialize_internal.h
@@ -319,6 +319,8 @@ void serialize_threads_shims(FILE *f, VirtualMachine *vm, Obj *prog);
 void serialize_type(FILE *f, SerializeContext *ctx, Type *ty);
 void serialize_type_decl(FILE *f, SerializeContext *ctx, Type *ty,
                          const char *name);
+void serialize_local_var_type_decl(FILE *f, SerializeContext *ctx, Type *ty,
+                                   const char *name);
 void serialize_type_defs_for_owner(FILE *f, SerializeContext *ctx,
                                    Obj *owner_fn);
 void serialize_uchar_shims(FILE *f, VirtualMachine *vm, Obj *prog);